chore: move coding tool to corral
