+++ tasks/catalyst/src/catalyst/tool_utils.py
@@ -1,182 +1,6 @@
 from pymatgen.core import Structure
 
 
-def generate_output_capture_code() -> str:
-    """
-    Generate the Python code string for capturing execution results.
-
-    This utility function creates the code that gets appended to user code
-    to capture variables and prepare them for JSON serialization.
-
-    Returns:
-        str: Python code string for variable capture
-    """
-    return """
-
-import json
-import types
-import sys
-from collections.abc import Iterable
-
-result_ = {}
-
-def is_json_serializable(obj, max_depth=10, current_depth=0):
-    '''Check if object is JSON serializable with depth limit'''
-    if current_depth > max_depth:
-        return False
-
-    try:
-        # Handle basic types
-        if obj is None or isinstance(obj, (bool, int, float, str)):
-            return True
-
-        # Handle sequences (but not strings)
-        if isinstance(obj, (list, tuple)):
-            return all(is_json_serializable(item, max_depth, current_depth + 1) for item in obj)
-
-        # Handle dictionaries
-        if isinstance(obj, dict):
-            return all(
-                isinstance(k, str) and is_json_serializable(v, max_depth, current_depth + 1)
-                for k, v in obj.items()
-            )
-
-        # Quick test with actual JSON serialization for edge cases
-        json.dumps(obj)
-        return True
-    except (TypeError, ValueError, RecursionError, OverflowError):
-        return False
-
-def safe_convert_to_serializable(obj):
-    '''Convert common non-serializable types to serializable ones'''
-    try:
-        import numpy as np
-        # Handle numpy arrays
-        if hasattr(obj, '__module__') and obj.__module__ == 'numpy':
-            if hasattr(obj, 'tolist'):
-                return obj.tolist()
-            elif hasattr(obj, 'item'):
-                return obj.item()
-    except ImportError:
-        pass
-
-    # Handle sets
-    if isinstance(obj, set):
-        return list(obj)
-
-    # Handle other iterables (but not strings/bytes)
-    if hasattr(obj, '__iter__') and not isinstance(obj, (str, bytes, dict)):
-        try:
-            return list(obj)
-        except:
-            pass
-
-    return obj
-
-# Try to capture from preferred variable names first
-preferred_vars = ['output', 'result', 'filtered_data', 'processed_data', 'dataset']
-captured = False
-
-for var_name in preferred_vars:
-    if var_name in locals():
-        var_value = locals()[var_name]
-        converted_value = safe_convert_to_serializable(var_value)
-        if is_json_serializable(converted_value):
-            result_[var_name] = converted_value
-            captured = True
-            break
-
-# Fallback: capture any user-defined variables
-if not captured:
-    excluded = {
-        '__builtins__', '__name__', '__doc__', '__package__', '__loader__',
-        '__spec__', '__annotations__', '__cached__', '__file__',
-        'json', 'types', 'sys', 'Iterable', 'result_', 'preferred_vars',
-        'captured', 'excluded', 'local_vars', 'is_json_serializable',
-        'safe_convert_to_serializable', 'var_name', 'var_value', 'converted_value'
-    }
-
-    local_vars = dict(locals())  # Create snapshot
-
-    # Collect all suitable variables
-    suitable_vars = {}
-    for var_name, var_value in local_vars.items():
-        if (not var_name.startswith('_') and
-            var_name not in excluded and
-            not isinstance(var_value, types.ModuleType) and
-            not callable(var_value)):
-
-            converted_value = safe_convert_to_serializable(var_value)
-            if is_json_serializable(converted_value):
-                suitable_vars[var_name] = converted_value
-
-    # If we have suitable variables, include them all
-    if suitable_vars:
-        result_.update(suitable_vars)
-
-print('EXECUTION_RESULT:', json.dumps(result_))
-"""
-
-
-def cast_timeout(timeout) -> int:
-    """
-    Safely convert timeout parameter to integer.
-
-    Args:
-        timeout: Timeout value (could be string, int, float, or None)
-
-    Returns:
-        int: Valid timeout value
-    """
-    if timeout is None:
-        return 300
-
-    try:
-        return int(float(timeout))  # Handle both "60" and "60.5"
-    except (ValueError, TypeError):
-        return 300  # Default fallback
-
-
-def ensure_directory_exists(file_path: str) -> None:
-    """
-    Ensure the parent directory of a file path exists.
-
-    Args:
-        file_path: Path to a file
-    """
-    from pathlib import Path
-
-    if file_path:
-        Path(file_path).parent.mkdir(parents=True, exist_ok=True)
-
-
-def parse_execution_output(stdout: str) -> tuple[dict, list[str]]:
-    """
-    Parse subprocess output to extract execution results and regular output.
-
-    Args:
-        stdout: Raw stdout from subprocess
-
-    Returns:
-        tuple: (execution_result dict, output_lines list)
-    """
-    import json
-    from contextlib import suppress
-
-    stdout_lines = stdout.strip().split("\n") if stdout.strip() else []
-    execution_result = {}
-    output_lines = []
-
-    for line in stdout_lines:
-        if line.startswith("EXECUTION_RESULT:"):
-            with suppress(json.JSONDecodeError):
-                execution_result = json.loads(line[17:])
-        else:
-            output_lines.append(line)
-
-    return execution_result, output_lines
-
-
 def load_structure(
     bulk_structure_path_or_string: str, from_path: bool = False
 ) -> Structure:
